fix: api 명세서대로 return 형식 수정

+++ backend/bizscanner/src/main/java/store/bizscanner/dto/response/salesResponse/QuarterSalesCountResponse.java
@@ -9,5 +9,5 @@ import java.util.List;
 @AllArgsConstructor
 public class QuarterSalesCountResponse {
 
-    private List<Integer> quarterSalesCount;
+    private List<Long> quarterSalesCount;
 }
